// src/lib/types/invoicing.ts
const SummaryDetailSchema = z.object({
  serie_numero: z.string(),
});

const SummaryDataSchema = z.object({
  id: z.number(),
  numero_completo: z.string(),
  fecha_resumen: z.string(),
  estado_proceso: z.string(),
  estado_sunat: z.string(),
  detalles: z.array(SummaryDetailSchema).optional().default([]),
});

export type SummaryData = z.infer<typeof SummaryDataSchema>;